ingest/requirements.txt
# Dependinte pentru pipeline-ul de ingest
# Install: pip install -r ingest/requirements.txt

# Conexiune Oracle in mod thin (NU necesita Oracle Instant Client)
oracledb==2.5.1

# Loading dataset HuggingFace
datasets==3.1.0

# Strip HTML din body-urile post-urilor
beautifulsoup4==4.12.3
lxml==5.3.0

# Bara de progress (optional, dar util pentru UX)
tqdm==4.67.1

# Helper pentru variabile .env (nu obligatoriu, dar curat)
python-dotenv==1.0.1

# Generare embeddings cu modelul all-MiniLM-L12-v2.
# Folosit pentru ca PAR-ul Oracle Object Storage cu modelul augmented a expirat.
# Native ARM64 pe Apple Silicon (PyTorch 2.x).
sentence-transformers==3.3.1
torch>=2.2.0
numpy>=1.26.0
